Added implemententation to Sim

+++ src/main/java/BlackHole.java
@@ -1,3 +1,3 @@
 public class BlackHole extends Environment{
-    public BlackHole (double height, double width, double wallElasticity) { super(99999999.99999, height, width, wallElasticity);}
+    public BlackHole (int height, int width, double wallElasticity) { super(99999999.99999, height, width, wallElasticity);}
 }
