+++ src/main/java/Earth.java
@@ -1,3 +1,3 @@
 public class Earth extends Environment {
-    public Earth(double height, double width, double wallElasticity) { super(9.81, height, width, wallElasticity);}
+    public Earth(int height, int width, double wallElasticity) { super(9.81, height, width, wallElasticity);}
 }
