+++ src/main/java/Jupiter.java
@@ -1,3 +1,3 @@
 public class Jupiter extends Environment {
-    public Jupiter(double height, double width, double wallElasticity) { super(24.79, height, width, wallElasticity);}
+    public Jupiter(int height, int width, double wallElasticity) { super(24.79, height, width, wallElasticity);}
 }
