+++ src/main/java/Mars.java
@@ -1,3 +1,3 @@
 public class Mars extends Environment {
-    public Mars(double height, double width, double wallElasticity) { super(3.71, height, width, wallElasticity);}
+    public Mars(int height, int width, double wallElasticity) { super(3.71, height, width, wallElasticity);}
 }
