+++ src/main/java/Moon.java
@@ -1,3 +1,3 @@
 public class Moon extends Environment {
-    public Moon (double height, double width, double wallElasticity) { super(1.62, height, width, wallElasticity);}
+    public Moon (int height, int width, double wallElasticity) { super(1.62, height, width, wallElasticity);}
 }
