+++ src/main/java/Neptune.java
@@ -1,3 +1,3 @@
 public class Neptune extends Environment {
-    public Neptune (double height, double width, double wallElasticity) { super(11.15, height, width, wallElasticity);}
+    public Neptune (int height, int width, double wallElasticity) { super(11.15, height, width, wallElasticity);}
 }
